# api_glewe/api_glewe/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": " OpenApi Documentation",
        "description": " Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/offer/user-validated-module": {
            "post": {
                "tags": [
                    "finish-module"
                ],
                "operationId": "11455660d5b0f6190d7b2af05decc10f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "module_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "string",
                                        "example": "99487892-f338-4740-88e7-b3377eafe173"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "I finished this module"
                    }
                }
            }
        }
    }
}